+++ packages/components/build.ts
@@ -6,4 +6,9 @@ defineBuild([
     outputFileName: 'auto-grid',
     vueComponent: true,
   },
+  {
+    entry: './src/infinite-scroll/index.ts',
+    outputFileName: 'infinite-scroll',
+    vueComponent: true,
+  },
 ]);