+++ packages/components/src/infinite-scroll/info.ts
@@ -1,0 +1,3 @@
+import type { Info } from '@/.vitepress/types/info';
+
+export const name: Info['name'] = '无限滚动';